refactor: 广告位模板和广告位管理代码优化（针对广告位编码字段)

+++ open-adexchange-model/src/main/java/top/openadexchange/model/SiteAdPlacement.java
@@ -58,7 +58,7 @@ public class SiteAdPlacement implements Serializable {
     /**
      * 底价
      */
-    private Double floorPrice;
+    private Long floorPrice;
     /**
      * 1-使用中, 0-禁用
      */
